Refactor game status form to have multiple pages
Visually broke up game status form into four 'pages' so users don't
need to unhide summary/details elements.
